+++ OsmParser/main.py
@@ -2,215 +2,13 @@
 import subprocess
 import sys
 
-from osmXMLparcer import *
-from osmGeometry import *
 from mdlMisc import *
-from vbFunctions import *
-from mdlSite import *
+from mdlOsmParser import readOsmXml
+from osmGeometry import clsOsmGeometry
+from mdlXmlParser import clsXMLparser
+from mdlSite import DoGeocodingForDatFile
 
 BUILD_PATH = 'd:\\_VFR_LANDMARKS_3D_RU'
-
-
-class T3DObject:
-    def __init__(self):
-        self.id = ""
-        self.type = ""
-        self.bbox = TBbox()
-        self.node_count = 0
-        self.NodeRefs = []
-        self.way_count = 0
-        self.WayRefs = []
-        self.name = ""
-        self.descr = ""
-        self.key_tags = ""
-        self.strWikipediaName = ""
-        self.tagBuilding = ""
-        self.tagArchitecture = ""
-        self.tagManMade = ""
-        self.tagBarrier = ""
-        self.tagTowerType = ""
-        self.tagAmenity = ""
-        self.tagDenomination = ""
-        self.tagStartDate = ""
-        self.tagRuins = ""
-        self.tagWikipedia = ""
-        self.tagAddrStreet = ""
-        self.tagAddrHouseNumber = ""
-        self.tagAddrCity = ""
-        self.tagAddrDistrict = ""
-        self.tagAddrRegion = ""
-        self.material = ""
-        self.colour = ""
-        self.dblHeight = 0
-        self.osmtags = {}
-        self.size = 0
-        self.blnHasBuildingParts = False
-
-        self.bbox.minLat = 0
-        self.bbox.minLon = 0
-        self.bbox.maxLat = 0
-        self.bbox.maxLon = 0
-
-    def getTag(self ,strKey):
-        return self.osmtags.get(strKey,'')
-
-    def isBuilding(self):
-        # if an object has building tag, it's probably a building (forget about building=no)
-        blnBuilding=(self.getTag("building") != "")
-        # Relation building is not really a building, it's just a group of building parts.
-        if (self.getTag("type") == "building"):
-            blnBuilding = False
-        return  blnBuilding
-
-    def isBuildingPart(self):
-        return (self.getTag("building:part") != "")
-
-
-
-# we will read osm file into a set of objects + complex structure with geometry
-def ReadOsmXml(strSrcOsmFile):
-
-    dblMaxHeight = 0
-    blnObjectIncomplete=False
-    print('Process start: list of building is generated')
-    objOsmGeom = clsOsmGeometry()
-    objXML = clsXMLparser()
-    Objects = []
-
-    objXML.OpenFile(strSrcOsmFile)
-    intModelsCreated = 0
-    while not objXML.bEOF:
-        objXML.ReadNextNode()
-        strTag = objXML.GetTag()
-        if strTag == 'node' or strTag == 'way' or strTag == 'relation':
-            osmObject = T3DObject()
-            osmObject.type = strTag
-            osmObject.id = objXML.GetAttribute('id')
-            blnObjectIncomplete= False
-
-        if strTag == 'node':
-            objOsmGeom.AddNode(osmObject.id, objXML.GetAttribute('lat'), objXML.GetAttribute('lon'))
-        # references to nodes in ways. we need to find coordinates
-        if strTag == 'nd':
-            node_id = objXML.GetAttribute('ref')
-            intNodeNo = objOsmGeom.FindNode(node_id)
-            if intNodeNo == - 1:
-                #raise Exception('FindNode', 'node not found! ' + node_id)
-                blnObjectIncomplete=True  
-            else: 
-                osmObject.NodeRefs.append( intNodeNo)
-                osmObject.node_count = osmObject.node_count + 1
-        # references to ways in relations. we need find coordinates
-        if strTag == 'member':
-            if objXML.GetAttribute('type') == 'way':
-                way_id = objXML.GetAttribute('ref')
-                intWayNo = objOsmGeom.FindWay(way_id)
-                if intWayNo == - 1:
-                    #raise Exception('FindWay', 'way not found! ' + way_id)
-                    blnObjectIncomplete=True  
-                else: 
-                    waybbox = objOsmGeom.GetWayBBox(intWayNo)
-                    osmObject.WayRefs.append([intWayNo, objXML.GetAttribute('role')])
-                    if osmObject.way_count == 0:
-                        osmObject.bbox.minLat = waybbox.minLat
-                        osmObject.bbox.minLon = waybbox.minLon
-                        osmObject.bbox.maxLat = waybbox.maxLat
-                        osmObject.bbox.maxLon = waybbox.maxLon
-                    else:
-                        if waybbox.minLat < osmObject.bbox.minLat:
-                            osmObject.bbox.minLat = waybbox.minLat
-                        if waybbox.maxLat > osmObject.bbox.maxLat:
-                            osmObject.bbox.maxLat = waybbox.maxLat
-                        if waybbox.minLon < osmObject.bbox.minLon:
-                            osmObject.bbox.minLon = waybbox.minLon
-                        if waybbox.maxLon > osmObject.bbox.maxLon:
-                            osmObject.bbox.maxLon = waybbox.maxLon
-                    osmObject.way_count = osmObject.way_count + 1
-
-        #get osmObject osm tags
-        if strTag == 'tag':
-            StrKey = objXML.GetAttribute('k')
-            strValue = objXML.GetAttribute('v')
-
-            osmObject.osmtags[StrKey] = strValue
-
-            if StrKey == 'height':
-                osmObject.dblHeight = ParseHeightValue(strValue)
-                if osmObject.dblHeight > dblMaxHeight:
-                    dblMaxHeight = osmObject.dblHeight
-            if StrKey == 'name':
-                osmObject.name = strValue
-            if StrKey == 'description':
-                osmObject.descr = strValue
-            if StrKey == 'building':
-                osmObject.tagBuilding = strValue
-                osmObject.key_tags = osmObject.key_tags + ' building=' + osmObject.tagBuilding
-            if StrKey == 'building:architecture':
-                osmObject.tagArchitecture = strValue
-            if StrKey == 'start_date':
-                osmObject.tagStartDate = strValue
-            if StrKey == 'man_made':
-                osmObject.tagManMade = strValue
-                osmObject.key_tags = osmObject.key_tags + ' man_made=' + osmObject.tagManMade
-            if StrKey == 'barrier':
-                osmObject.tagBarrier = strValue
-                osmObject.key_tags = osmObject.key_tags + ' barrier=' + osmObject.tagBarrier
-            # wikipedia
-            if StrKey == 'wikipedia':
-                osmObject.tagWikipedia = strValue
-            if StrKey == 'addr:street':
-                osmObject.tagAddrStreet = strValue
-            if StrKey == 'addr:housenumber':
-                osmObject.tagAddrHouseNumber = strValue
-            if StrKey == 'addr:city':
-                osmObject.tagAddrCity = strValue
-            if StrKey == 'addr:district':
-                osmObject.tagAddrDistrict = strValue
-            if StrKey == 'addr:region':
-                osmObject.tagAddrRegion = strValue
-            #ref_temples_ru
-            if StrKey == 'ref:temples.ru':
-                ref_temples_ru = strValue
-                #print ref_temples_ru
-            if StrKey == 'amenity':
-                osmObject.tagAmenity = strValue
-            if StrKey == 'denomination':
-                osmObject.tagDenomination = strValue
-            if StrKey == 'tower:type':
-                osmObject.tagTowerType = strValue
-            if StrKey == 'building:material':
-                osmObject.material = strValue
-            #for buildings we have building:colour, for other objects, e.g. fences, just colour
-            if ( StrKey == 'building:colour' )  or  ( StrKey == 'colour' ) :
-                osmObject.colour = strValue
-                if Left(strValue, 1) == '#':
-                    osmObject.colour = GetColourName(strValue)
-            if StrKey == 'ruins':
-                osmObject.tagRuins = strValue
-
-
-        if strTag == '/way':
-            intWayNo = objOsmGeom.AddWay(osmObject.id, osmObject.NodeRefs, osmObject.node_count)
-            osmObject.bbox = objOsmGeom.GetWayBBox(intWayNo)
-            osmObject.size = objOsmGeom.CalculateWaySize(intWayNo)
-
-        if strTag == '/relation':
-            osmObject.size = objOsmGeom.CalculateRelationSize(osmObject.WayRefs, osmObject.way_count)
-            # bbox is already calculated above
-
-        # Closing node
-        if strTag == '/node' or strTag == '/way' or strTag == '/relation':
-            if (blnObjectIncomplete != True) and (osmObject.type != 'node'):
-                # we will return only completed objects, and we will skip nodes (to save CPU time)
-                Objects.append(osmObject)
-            else:
-                #print('Object is incomplete ' + osmObject.type +' ' +  osmObject.id)
-                pass
-
-    objXML.CloseFile()
-
-    return objOsmGeom, Objects
-
 
 def processBuildings(objOsmGeom, Objects, strQuadrantName, strObjectsWithPartsFileName, strOutputFile, OSM_3D_MODELS_PATH):
     j = 0
@@ -229,13 +27,15 @@ def processBuildings(objOsmGeom, Objects, strQuadrantName, strObjectsWithPartsFi
         if blnBuilding or blnFence:
             SelectedObjects.append(osmObject)
         else:
-            print( "not a building: " + osmObject.type + " " + osmObject.id)
-            # pass
+            #print( "not a building: " + osmObject.type + " " + osmObject.id)
+            pass
 
     fo = open(strOutputFile, 'w', encoding="utf-8")
     for osmObject in SelectedObjects:
         heightbyparts = 0
         numberofparts = 0
+        strHeight = osmObject.getTag('height')
+        osmObject.dblHeight = parseHeightValue(strHeight)
         blnFence = (osmObject.tagBarrier == 'fence') or (osmObject.tagBarrier == 'wall')
 
         # Rewrite osmObject as osm file!
@@ -255,19 +55,24 @@ def processBuildings(objOsmGeom, Objects, strQuadrantName, strObjectsWithPartsFi
                                                 osmObject.tagAmenity, osmObject.getTag('religion'),
                                                 osmObject.tagDenomination, osmObject.tagBarrier, osmObject.size,
                                                 osmObject.tagRuins)
+
+
+
+
         ref_temples_ru = osmObject.getTag('ref:temples.ru')
         j = j + 1
-        fo.write(str(j) + '|' + osmObject.type + '|' + osmObject.id + '|' + str(osmObject.bbox.minLat) + '|' + str(
-            osmObject.bbox.minLon) + '|' +
-                 str(osmObject.bbox.maxLat) + '|' + str(
-            osmObject.bbox.maxLon) + '|' + osmObject.name + '|' + osmObject.descr + '|' + ref_temples_ru + '|' +
-                 strBuildingType + '|' + str(Round(osmObject.size)) + '|' + str(
-            Round(osmObject.dblHeight)) + '|' + osmObject.colour + '|' + osmObject.material + '|' +
-                 GuessBuildingStyle(osmObject.tagArchitecture, osmObject.tagStartDate) + '|' + ParseStartDateValue(
-            osmObject.tagStartDate) + '|' + osmObject.tagWikipedia + '|' +
-                 osmObject.tagAddrStreet + '|' + osmObject.tagAddrHouseNumber + '|' + osmObject.tagAddrCity + '|' + osmObject.tagAddrDistrict + '|' + osmObject.tagAddrRegion + '|' +
-                 str(osmObject.blnHasBuildingParts and (osmObject.dblHeight > 0)) + '|' + str(numberofparts) +
-                 '\n')
+        fo.write(str(j) + '|' + osmObject.type + '|' + osmObject.id + '|' + str(osmObject.bbox.minLat) + '|'
+                 + str(osmObject.bbox.minLon) + '|'
+                 + str(osmObject.bbox.maxLat) + '|'
+                 + str(osmObject.bbox.maxLon) + '|' + osmObject.name + '|' + osmObject.descr + '|' + ref_temples_ru + '|'
+                 + strBuildingType + '|' + str(Round(osmObject.size)) + '|'
+                 + str(Round(osmObject.dblHeight)) + '|' + osmObject.colour + '|' + osmObject.material + '|'
+                 + GuessBuildingStyle(osmObject.tagArchitecture, osmObject.tagStartDate) + '|'
+                 + parseStartDateValue( osmObject.tagStartDate) + '|' + osmObject.tagWikipedia + '|'
+                 + osmObject.tagAddrStreet + '|' + osmObject.tagAddrHouseNumber + '|' + osmObject.tagAddrCity + '|'
+                 + osmObject.tagAddrDistrict + '|' + osmObject.tagAddrRegion + '|'
+                 + str(osmObject.blnHasBuildingParts and (osmObject.dblHeight > 0)) + '|' + str(numberofparts)
+                 + '\n')
     fo.close()
 
     # miracle: update totals file
@@ -284,7 +89,7 @@ def processBuildings(objOsmGeom, Objects, strQuadrantName, strObjectsWithPartsFi
     print(str(j) + ' objects detected, ' + str(intModelsCreated) + ' 3d models created')
 
 
-def ParseHeightValue(str):
+def parseHeightValue(str):
     if Right(str, 2) == ' м':
         str = Left(str, Len(str) - 2)
     if Right(str, 2) == ' m':
@@ -293,7 +98,9 @@ def ParseHeightValue(str):
         str = '0'
     if Right(str,1) == "'":
         str=Trim(Left(str, Len(str) - 1))
-        str=float(str)*0.3048 
+        str=float(str)*0.3048
+    if str == "":
+        str = '0'
     if not IsNumeric(str):
         print ("Unparsed height value: " + str)
         str = '0'
@@ -301,7 +108,7 @@ def ParseHeightValue(str):
     return fn_return_value
 
 
-def ParseStartDateValue(strDate):
+def parseStartDateValue(strDate):
 
     strResult = ""
     strModifier = ""
@@ -374,7 +181,7 @@ def GuessBuildingStyle(strArchitecture, strDate):
         strResult = strArchitecture
     else:
         if strDate != '':
-            strDate = ParseStartDateValue(strDate)
+            strDate = parseStartDateValue(strDate)
             if strDate < '1250':
                 strResult = '~pre-mongolian'
             elif strDate < '1650':
@@ -520,7 +327,7 @@ def RewriteOsmFile(object1, strObjectsWithPartsFileName, OSM_3D_MODELS_PATH):
             if StrKey == 'building:part' and strValue != 'no':
                 obj_is_building_part = True
             if StrKey == 'height':
-                obj_height = ParseHeightValue(strValue)
+                obj_height = parseHeightValue(strValue)
             if StrKey == 'building:levels':
                 obj_levels = strValue
                  
@@ -637,7 +444,7 @@ def ProcessQuadrant(strQuadrantName):
 
     subprocess.call(BUILD_PATH + '\\cleanup.bat', cwd=strWorkingFolder + '\\osm_3dmodels')
 
-    objOsmGeom, Objects = ReadOsmXml(strWorkingFolder + '\\osm_data\\objects-all.osm')
+    objOsmGeom, Objects = readOsmXml(strWorkingFolder + '\\osm_data\\objects-all.osm')
     processBuildings(objOsmGeom, Objects, strQuadrantName,
                      strWorkingFolder + '\\osm_data\\objects-with-parts.osm',
                      strWorkingFolder + '\\' + strQuadrantName + '.dat',
