+++ OsmParser/planner.py
@@ -1,4 +1,4 @@
-from osmXMLparcer import *
+from mdlXmlParser import *
 from osmGeometry import *
 from vbFunctions import *
 from mdlDBMetadata import *
